Move 'utils' regions from `settings`/`build` Gradle files to `buildSrc`
