+++ settings.gradle.kts
@@ -40,34 +40,3 @@ defaultRepositories {
 }
 
 enableFeaturePreview("GRADLE_METADATA")
-
-//region utils
-
-fun gradleWrapper(version: String, distribution: Wrapper.DistributionType) {
-    gradle.rootProject {
-        afterEvaluate {
-            tasks.named<Wrapper>("wrapper") {
-                gradleVersion = version
-                distributionType = distribution
-            }
-        }
-    }
-}
-
-typealias Version = String
-
-operator fun Version.invoke(vararg plugins: PluginDependencySpec) =
-    plugins.forEach { it version this }
-
-fun defaultRepositories(repositories: RepositoryHandler.() -> Unit) {
-    pluginManagement {
-        this.repositories.repositories()
-        this.repositories.gradlePluginPortal()
-    }
-    gradle.allprojects {
-        this.buildscript.repositories.repositories()
-        this.repositories.repositories()
-    }
-}
-
-//endregion
